% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{demoevents}
\alias{demoevents}
\title{Create demo date events}
\usage{
demoevents()
}
\value{
Data which is set to be equal to datetimes roughly around the users current system time
}
\description{
Create demo date events
}
\examples{
\donttest{
demodates()
}
}
