% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{demoevents}
\alias{demoevents}
\alias{randomevents}
\title{Create demo date events}
\usage{
demoevents()

randomevents()
}
\value{
Data which is capable of being used in fullercalendar, with datetimes roughly around the users current system time
}
\description{
Create demo date events
}
\examples{
demoevents()
randomevents()
}
