% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{isodate}
\alias{isodate}
\title{Turn Date object to a formatted date needed for the FullCalendar js to recognise}
\usage{
isodate(date)
}
\arguments{
\item{date}{A valid Date object}
}
\value{
a string of the date
}
\description{
Turn Date object to a formatted date needed for the FullCalendar js to recognise
}
\examples{
isodate(Sys.Date())
}
