% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fullercalendar.R
\name{fullercalendar}
\alias{fullercalendar}
\title{<Add Title>}
\usage{
fullercalendar(events = NULL, settings = list(), width = NULL,
  height = NULL, elementId = NULL)
}
\description{
<Add Description>
}
