% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fullercalendar.R
\name{fullercalendar}
\alias{fullercalendar}
\title{Display a FullCalendar widget}
\usage{
fullercalendar(events = list(), ..., plugins = c("interaction",
  "dayGrid", "timeGrid"), width = NULL, height = NULL,
  elementId = NULL, defaultView = "dayGridMonth", defaultDate = NULL)
}
\arguments{
\item{events}{a data frame with one row per event where the columns needs
to follow the naming conventions by FullCalendar shown here: https://fullcalendar.io/docs/event-object.
This data frame will then be translated into a json array by means of \code{jsonlite::toJSON}. The required
columns are \code{title}, a string, and \code{start} as string parse-able as a date, for example, "2017-02-28".}

\item{...}{Other options which can be set specifically, that are not specified here: https://fullcalendar.io/docs/dynamic-options. See details.}

\item{plugins}{List of plugins to use when initialising the calendar}

\item{width}{Fixed width for widget (in css units). The default is NULL, which results in intelligent automatic sizing based on the widget's container.}

\item{height}{Fixed height for widget (in css units). The default is NULL, which results in intelligent automatic sizing based on the widget's container.}

\item{elementId}{Use an explicit element ID for the widget (rather than an automatically generated one). Useful if you have other JavaScript that needs to explicitly discover and interact with a specific widget instance.}

\item{defaultView}{A name of any of the available views, such as 'dayGridWeek', 'timeGridDay', 'listWeek'}

\item{defaultDate}{The initial date displayed when the calendar first loads. Date When not specified, this value defaults to the current date. This value can be anything that can parse into a Date, including an ISO8601 date string like "2014-02-01".}
}
\description{
This function creates a FullCalendar htmlwidget that will show up
in the viewer pane, if you're running RStudio, or in your Rmarkdown
document.
}
\details{
The page https://fullcalendar.io/docs#toc defines the many options available to modifying the calendar object. Currently, most non-method properties are implemented.
}
\examples{
fullercalendar(events =  demoevents())

\donttest{
fullercalendar(demoevents(),
               defaultView = "timeGridWeek",
               weekends = TRUE,
               hiddenDays= c(2, 6),
               slotDuration = '00:10:00',
               firstDay = 1,
               weekNumbers = TRUE,
               selectable = TRUE,
               selectMirror = TRUE,
               editable = TRUE,
               dropable = TRUE,

)
}

}
